Reduce dependency on Ranklib
- Refactor the lucene query LtrQuery into RankerQuery
- Introduces a Feature interface
- Make RankerQuery ranker agnostic with the LtrRanker interface
- Added a simple LinearRanker
- Make RankerQuery a bit more lucene compliant (still need to figure out the
  impact on existing models that uses a Similarity that implements queryNorm)
- LtrQueryQuery now creates a RankingQuery with RanklibRanker

This is just a suggestion and packages are maybe not appropriate but Ranklib
dependent classes are now in their own package.

The purpose is to start to reduce the dependency on Ranklib so that users
can use their own training libraries such as xgboost.

It's still unclear how to refactor LtrQueryBuilder so that it remains back
compatible with existing clients but still allow to use non-ranklib models.

A regression tree implementation is needed.
The format to build a LinearModel or a regression tree model will have to be defined.